DataManagement page, break sort out into dedicated hook
