+++ client/components/DataManagement/hooks.js
@@ -1,0 +1,79 @@
+import { useMemo, useState } from 'react';
+import {
+    DATA_MANAGEMENT_TABLE_SORT_OPTIONS,
+    TABLE_SORT_ORDERS
+} from '../../utils/constants';
+
+export const useDataManagementTableSorting = (
+    testPlans,
+    testPlanVersions,
+    ats
+) => {
+    const [activeSort, setActiveSort] = useState({
+        key: DATA_MANAGEMENT_TABLE_SORT_OPTIONS.PHASE,
+        direction: TABLE_SORT_ORDERS.ASC
+    });
+
+    const sortedTestPlans = useMemo(() => {
+        const phaseOrder = {
+            RD: 0,
+            DRAFT: 1,
+            CANDIDATE: 2,
+            RECOMMENDED: 3
+        };
+        const directionMod =
+            activeSort.direction === TABLE_SORT_ORDERS.ASC ? -1 : 1;
+
+        const getTestPlanVersionOverallPhase = t => {
+            return (
+                Object.keys(phaseOrder).find(phaseKey =>
+                    testPlanVersions.some(
+                        ({ phase, testPlan }) =>
+                            testPlan.directory === t.directory &&
+                            phase === phaseKey
+                    )
+                ) || 'RD'
+            );
+        };
+
+        const sortByName = (a, b, dir = directionMod) =>
+            dir * (a.title < b.title ? -1 : 1);
+
+        const sortByAts = (a, b) => {
+            const countA = ats.length; // Stubs based on current rendering in DataManagementRow
+            const countB = ats.length;
+            if (countA === countB) return sortByName(a, b);
+            return directionMod * (countA - countB);
+        };
+
+        const sortByPhase = (a, b) => {
+            const testPlanVersionOverallA = getTestPlanVersionOverallPhase(a);
+            const testPlanVersionOverallB = getTestPlanVersionOverallPhase(b);
+            if (testPlanVersionOverallA === testPlanVersionOverallB)
+                return sortByName(a, b, 1);
+            return (
+                directionMod *
+                (phaseOrder[testPlanVersionOverallA] -
+                    phaseOrder[testPlanVersionOverallB])
+            );
+        };
+
+        const sortFunctions = {
+            NAME: sortByName,
+            ATS: sortByAts,
+            PHASE: sortByPhase
+        };
+
+        return testPlans.slice().sort(sortFunctions[activeSort.key]);
+    }, [activeSort, testPlans]);
+
+    const updateSort = ({ key, direction }) => {
+        setActiveSort({ key, direction });
+    };
+
+    return {
+        sortedTestPlans,
+        updateSort,
+        activeSort
+    };
+};
